Split out Filesystem classes to their own files
